pyproject.toml
[project]
name = "extremoambiente-agent"
version = "0.2.0"
description = "AI assistant for Extremo Ambiente — adventure tourism event planning"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.3.0",
    "python-dotenv>=1.0.0",
    "langgraph-cli[inmem]>=0.4.18",
]

[tool.setuptools.packages.find]
include = ["agent*"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
